<fix>[host]: move network interface type api to promethus
Resolves: ZSTAC-64822

(cherry picked from commit 68cae512f838658d533745cf1905673ae6621749)
